Address review comments on introducing a new interface for Http source configuration sharing between different http based sources
