+++ data-prepper-plugins/http-source/src/main/java/org/opensearch/dataprepper/plugins/source/loghttp/HTTPSource.java
@@ -13,6 +13,7 @@ import com.linecorp.armeria.server.healthcheck.HealthCheckService;
 import com.linecorp.armeria.server.throttling.ThrottlingService;
 import org.opensearch.dataprepper.HttpRequestExceptionHandler;
 import org.opensearch.dataprepper.armeria.authentication.ArmeriaHttpAuthenticationProvider;
+import org.opensearch.dataprepper.http.HttpServerConfig;
 import org.opensearch.dataprepper.http.LogThrottlingRejectHandler;
 import org.opensearch.dataprepper.http.LogThrottlingStrategy;
 import org.opensearch.dataprepper.metrics.PluginMetrics;
@@ -51,7 +52,7 @@ public class HTTPSource implements Source<Record<Log>> {
     public static final String REGEX_HEALTH = "regex:^/(?!health$).*$";
     static final String SERVER_CONNECTIONS = "serverConnections";
 
-    private final HTTPSourceConfig sourceConfig;
+    private final HttpServerConfig sourceConfig;
     private final CertificateProviderFactory certificateProviderFactory;
     private final ArmeriaHttpAuthenticationProvider authenticationProvider;
     private final HttpRequestExceptionHandler httpRequestExceptionHandler;
